Fleshed out some classes [Needs work ]
I filled out some of the classes including parent object, lemming and
object manager.
Wrote comments for things which could be done.
Have lemming on screen basic way for now but I will change this when
manager and factory are built properly.
Will work on this more in the morning before meeting.