Made entity cycle work kind of
I now have a lemming appear on the screen through the object manager and
object factory classes. It falls down the screen through the use of it's
"Default" mind. I am trying now to make the factory create a second
lemming after a few seconds of creating the first.
Taking a small break for now. More commits to follow.