Tried to fix fact timer
I tried to fix the timer in the factory to make it so that lemmings are
only created when delay is set to false.
I realised however that it needs to be timed somewhere else or at least
have lemmings created last, after all the other elements of a level e.g.
spider, floor etc.
Or we could create them all and have them just be drawn every 5 seconds?
I also put in comments where I thought things needed to be added or
changed later.