Updated draw method
I updated the draw method, in parent object, to include all of the
necessary parameters e.g. sprite effects. Because of this I needed to
include those properties in all of the objects.
I made it so that the edges images flip properly now and there is more
of a border around the cave/screen.
Spider will also be able to rotate now when we get to that.