Made SceneManager work[Without SceneGraph]
I have the scene manager working with objectmanager and renderer. It can
now also still work if an entity needs to be removed from the scene,
like when a lemming needs to terminate.
Added comments describing what each section does in the SM as well.
Need to look up what the SceneGraph actually does so that it can be
implemented.